refactor fhir_receiver_processor to extract fhir_receiver_processor_spark.py
